This is the original source code
It is downloaded from the Nikhef website
and created by Jeroen van Leerdam. It is
working, except for the known issues.

The required LabVIEW version is the name
of this branch (LabVIEW 8.6) editing it
in a higher version is possible, but it
breaks the backwards compatibility, so
only do this is you must. When you have
to edit it in a higher version, it's
better to create a new branch with the
LabVIEW version as the branch name, i.e.: LabVIEW 10.0