Fixed issue #4
The VI where the data is written to the MySQL database
now checks the result of the query. If this is an error
corresponding to the error code of a ODBC connection
error, the VI will set the MySQL connection status
to disconnected. When the MySQL database is back up,
the connection will be re-established and the data
backed-up in the queues will be written to the database.